// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const StatusPeminjaman = {
  direservasi: 'direservasi',
  dibatalkan: 'dibatalkan',
  dipinjam: 'dipinjam',
  dikembalikan: 'dikembalikan',
  terlambat: 'terlambat'
} as const

export type StatusPeminjaman = (typeof StatusPeminjaman)[keyof typeof StatusPeminjaman]


export const StatusBayar = {
  belum_bayar: 'belum_bayar',
  sudah_bayar: 'sudah_bayar'
} as const

export type StatusBayar = (typeof StatusBayar)[keyof typeof StatusBayar]


export const Role = {
  CUSTOMER: 'CUSTOMER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const JK = {
  L: 'L',
  P: 'P'
} as const

export type JK = (typeof JK)[keyof typeof JK]


export const keteranganDenda = {
  tidak_ada: 'tidak_ada',
  sobek: 'sobek',
  noda: 'noda',
  rusak_total: 'rusak_total',
  kehilangan_buku: 'kehilangan_buku'
} as const

export type keteranganDenda = (typeof keteranganDenda)[keyof typeof keteranganDenda]